Moving directory to handwired, which more accurately describes the project
