+++ keyboards/handwired/unicomp_mini_m/unicomp_mini_m.c
@@ -13,4 +13,4 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-#include "mini_m.h"
+#include "unicomp_mini_m.h"
